+++ gneiss/plot/__init__.py
@@ -26,7 +26,8 @@ Functions
 from ._heatmap import heatmap
 from ._blobtree import diamondtree
 from ._radial import radialplot
-from ._plot import ols_summary
+from ._plot import ols_summary, lme_summary, dendrogram_heatmap
 
 
-__all__ = ["heatmap", "diamondtree", "radialplot", "ols_summary"]
+__all__ = ["heatmap", "diamondtree", "radialplot",
+           "ols_summary", "lme_summary", "dendrogram_heatmap"]