+++ gneiss/plot/_plot.py
@@ -7,13 +7,21 @@
 # ----------------------------------------------------------------------------
 import os
 import pandas as pd
+from skbio import TreeNode
+from skbio.stats.composition import clr, centralize
 
-from qiime2.plugin import Int
 from gneiss.plugin_setup import plugin
 from gneiss.plot._radial import radialplot
+from gneiss.plot._heatmap import heatmap
 
 from gneiss.regression._ols import OLSModel
-from gneiss.regression._type import LinearRegression_g
+from gneiss.regression._mixedlm import LMEModel
+from gneiss.regression._type import (LinearRegression_g,
+                                     LinearMixedEffects_g)
+from q2_types.tree import Phylogeny, Rooted
+from q2_composition.plugin_setup import Composition
+from q2_types.feature_table import FeatureTable
+from qiime2.plugin import Int, MetadataCategory
 
 try:
     from bokeh.embed import file_html
@@ -86,8 +94,62 @@ def _projected_residuals(model):
     return p
 
 
-def ols_summary(output_dir: str, model: OLSModel, ndim=10) -> None:
+def _deposit_results(model, output_dir):
+    """ Store all of the core regression results into a folder. """
+    pred = model.predict()
+    pred.to_csv(os.path.join(output_dir, 'predicted.csv'),
+                header=True, index=True)
+    coefficients = model.coefficients()
+    coefficients.to_csv(os.path.join(output_dir, 'coefficients.csv'),
+                        header=True, index=True)
+    residuals = model.residuals()
+    residuals.to_csv(os.path.join(output_dir, 'residuals.csv'),
+                     header=True, index=True)
+    predicted = model.predict()
+    predicted.to_csv(os.path.join(output_dir, 'predicted.csv'),
+                     header=True, index=True)
+    pvalues = model.pvalues
+    pvalues.to_csv(os.path.join(output_dir, 'pvalues.csv'),
+                   header=True, index=True)
+    balances = model.balances
+    balances.to_csv(os.path.join(output_dir, 'balances.csv'),
+                    header=True, index=True)
 
+
+def _deposit_results_html(index_f):
+    """ Create links to all of the regression results. """
+    index_f.write(('<th>Coefficients</th>\n'))
+    index_f.write(('<a href="coefficients.csv">'
+                   'Download as CSV</a><br>\n'))
+    index_f.write(('<th>Coefficient pvalues</th>\n'))
+    index_f.write(('<a href="pvalues.csv">'
+                   'Download as CSV</a><br>\n'))
+    index_f.write(('<th>Raw Balances</th>\n'))
+    index_f.write(('<a href="balances.csv.csv">'
+                   'Download as CSV</a><br>\n'))
+    index_f.write(('<th>Predicted Proportions</th>\n'))
+    index_f.write(('<a href="predicted.csv">'
+                   'Download as CSV</a><br>\n'))
+    index_f.write(('<th>Residuals</th>\n'))
+    index_f.write(('<a href="residuals.csv">'
+                   'Download as CSV</a><br>\n'))
+
+
+# OLS summary
+def ols_summary(output_dir: str, model: OLSModel, ndim=10) -> None:
+    """ Summarizes the ordinary least squares fit.
+
+    Parameters
+    ----------
+    output_dir : str
+        Directory where all of the regression results and
+        summaries will be stored.
+    model : OLSModel
+        Ordinary Least Squares model that contains the model fit and the
+        regression results.
+    ndim : int
+        The number of dimensions to summarize.
+    """
     # Cross validation
     cv = model.loo()
     # Relative importance of explanatory variables
@@ -120,25 +182,7 @@ def ols_summary(output_dir: str, model: OLSModel, ndim=10) -> None:
 
     p23 = hplot(p2, p3)
 
-    # Deposit all regression results
-    pred = model.predict()
-    pred.to_csv(os.path.join(output_dir, 'predicted.csv'),
-                header=True, index=True)
-    coefficients = model.coefficients()
-    coefficients.to_csv(os.path.join(output_dir, 'coefficients.csv'),
-                        header=True, index=True)
-    residuals = model.residuals()
-    residuals.to_csv(os.path.join(output_dir, 'residuals.csv'),
-                     header=True, index=True)
-    predicted = model.predict()
-    predicted.to_csv(os.path.join(output_dir, 'predicted.csv'),
-                     header=True, index=True)
-    pvalues = model.pvalues
-    pvalues.to_csv(os.path.join(output_dir, 'pvalues.csv'),
-                   header=True, index=True)
-    balances = model.balances
-    balances.to_csv(os.path.join(output_dir, 'balances.csv'),
-                    header=True, index=True)
+    _deposit_results(model, output_dir)
 
     index_fp = os.path.join(output_dir, 'index.html')
     with open(index_fp, 'w') as index_f:
@@ -149,21 +193,7 @@ def ols_summary(output_dir: str, model: OLSModel, ndim=10) -> None:
         index_f.write(relimp.to_html())
         index_f.write('<th>Cross Validation</th>')
         index_f.write(cv.to_html())
-        index_f.write(('<th>Coefficients</th>\n'))
-        index_f.write(('<a href="coefficients.csv">'
-                       'Download as CSV</a><br>\n'))
-        index_f.write(('<th>Coefficient pvalues</th>\n'))
-        index_f.write(('<a href="pvalues.csv">'
-                       'Download as CSV</a><br>\n'))
-        index_f.write(('<th>Raw Balances</th>\n'))
-        index_f.write(('<a href="balances.csv.csv">'
-                       'Download as CSV</a><br>\n'))
-        index_f.write(('<th>Predicted Proportions</th>\n'))
-        index_f.write(('<a href="predicted.csv">'
-                       'Download as CSV</a><br>\n'))
-        index_f.write(('<th>Residuals</th>\n'))
-        index_f.write(('<a href="residuals.csv">'
-                       'Download as CSV</a><br>\n'))
+        _deposit_results_html(index_f)
         ess_tree_html = file_html(p1, CDN, 'Explained Sum of Squares')
         index_f.write(ess_tree_html)
         reg_smry_html = file_html(p23, CDN, 'Prediction and Residual plot')
@@ -186,4 +216,136 @@ plugin.visualizers.register_function(
                  "explained sum of squares, coefficients, "
                  "coefficient pvalues, coefficient of determination, "
                  "predicted fit, and residuals")
+)
+
+
+# LME summary
+def lme_summary(output_dir: str, model: LMEModel, ndim=10) -> None:
+    """ Summarizes the ordinary linear mixed effects model.
+
+    Parameters
+    ----------
+    output_dir : str
+        Directory where all of the regression results and
+        summaries will be stored.
+    model : LMEModel
+        Linear Mixed Effects model that contains the model fit and the
+        regression results.
+    ndim : int
+        The number of dimensions to summarize.
+    """
+    # log likelihood
+    loglike = pd.Series({r.model.endog_names: r.model.loglike(r.params)
+                         for r in model.results})
+
+    # Summary object
+    smry = model.summary(ndim=10)
+
+    t = model.tree
+    for i, n in enumerate(t.postorder()):
+        n.size = 10
+        if n.is_root():
+            n.size = 20
+        elif n.name == n.parent.children[0].name:
+            n.color = '#00FF00'  # left child is green
+        else:
+            n.color = '#FF0000'  # right child is red
+
+        if not n.is_tip():
+            t.length = -loglike.loc[n.name]
+
+    p1 = radialplot(t, node_color='color', figsize=(800, 800))
+    p1.title.text = 'Loglikelihood of submodels'
+    p1.title_location = 'above'
+    p1.title.align = 'center'
+    p1.title.text_font_size = '18pt'
+
+    # 2D scatter plot for prediction on PB
+    p2 = _projected_prediction(model)
+    p3 = _projected_residuals(model)
+
+    p23 = hplot(p2, p3)
+
+    # Deposit all regression results
+    _deposit_results(model, output_dir)
+
+    index_fp = os.path.join(output_dir, 'index.html')
+    with open(index_fp, 'w') as index_f:
+        index_f.write('<html><body>\n')
+        index_f.write('<h1>Simplicial Linear Mixed Effects Summary</h1>\n')
+        index_f.write(smry.as_html())
+        _deposit_results_html(index_f)
+        ess_tree_html = file_html(p1, CDN, 'Loglikelihood')
+        index_f.write(ess_tree_html)
+        reg_smry_html = file_html(p23, CDN, 'Prediction and Residual plot')
+        index_f.write(reg_smry_html)
+        index_f.write('</body></html>\n')
+
+
+plugin.visualizers.register_function(
+    function=lme_summary,
+    inputs={'model': LinearMixedEffects_g},
+    parameters={'ndim': Int},
+    input_descriptions={
+        'model': 'The fitted simplicial ordinary least squares model.'
+    },
+    parameter_descriptions={
+        'ndim': 'Number of dimensions to summarize.'},
+    name='Simplicial Linear Mixed Effects Summary plots.',
+    description=("Visualize the summary statistics of a simplicial "
+                 "linear mixed effectsmodel. This includes the "
+                 "loglikhood estimates, coefficients, "
+                 "coefficient pvalues, coefficient of determination, "
+                 "predicted fit and residuals")
+)
+
+
+# Heatmap
+def dendrogram_heatmap(output_dir: str, table: pd.DataFrame,
+                       tree: TreeNode, metadata: MetadataCategory,
+                       ndim=10):
+    nodes = [n.name for n in tree.levelorder()]
+    nlen = min(ndim, len(nodes))
+    highlights = pd.DataFrame([['#00FF00', '#FF0000']] * nlen,
+                              index=nodes[:nlen])
+
+    mat = pd.DataFrame(clr(centralize(table)),
+                       index=table.index,
+                       columns=table.columns)
+    # TODO: There is a bit of hard-coded constants here
+    # will need to have some adaptive defaults set in the future
+    fig = heatmap(mat, tree, metadata.to_series(), highlights,
+                  highlight_width=0.01, dendrogram_width=50,
+                  figsize=(12, 8))
+    fig.savefig(os.path.join(output_dir, 'heatmap.svg'))
+
+    index_fp = os.path.join(output_dir, 'index.html')
+    with open(index_fp, 'w') as index_f:
+        index_f.write('<html><body>\n')
+        index_f.write('<h1>Dendrogram heatmap</h1>\n')
+        index_f.write('<img src="heatmap.svg" alt="heatmap">')
+        index_f.write('</body></html>\n')
+
+
+plugin.visualizers.register_function(
+    function=dendrogram_heatmap,
+    inputs={'table': FeatureTable[Composition],
+            'tree': Phylogeny[Rooted]},
+    parameters={'metadata': MetadataCategory, 'ndim': Int},
+    input_descriptions={
+        'table': ('The feature table that will be plotted as a heatmap. '
+                  'This table is assumed to have strictly positive values.'),
+        'tree': ('A hierarchy of feature identifiers where each tip'
+                 'corresponds to the feature identifiers in the table. '
+                 'This tree can contain tip ids that are not present in '
+                 'the table, but all feature ids in the table must be '
+                 'present in this tree.')},
+    parameter_descriptions={
+        'metadata': ('Metadata to group the samples. '),
+        'ndim': 'Number of dimensions to highlight.'},
+    name='Dendrogram heatmap.',
+    description=("Visualize the feature tables as a heatmap. "
+                 "with samples sorted along a specified metadata category "
+                 "and features clustered together specified by the tree."
+                 "In addition, the heatmap values are clr transformed.")
 )